+++ src/lib/internal/validators.ts
@@ -1,0 +1,8 @@
+/**
+ * @ignore
+ */
+const isFunction = (fn: any) => typeof fn === 'function';
+
+export {
+    isFunction
+};