+++ CHANGELOG.md
@@ -8,3 +8,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - `toJson` function
  - Markdown docs generation using typedoc
+ - `absoluteUrl` function